changing button/gridbutton functions in usercode oop style looks similar to nodejs (or atleast discordjs) were even "huge" data in functions wouldn't be write heavy also helps for usercode to also use the data further than just the function alone since it allows splicing the information easier.
Also sge2 refactor was 100% unnecessary it's goal was trying to simplify developing in sge1 engine side and was becoming just more complicated to work with so plans for just taking best parts of sge2 and merge them to sge1 and remove sge2.

sge refactor will require all usercode to change header location for future in tl\ph\sge folder instead of tl\sge\ which will remain stationary for the rest of the development. Will also help usercode to use other parts of the library if wanted since all public headers will be in tl\ph